+++ packages/shared/src/index.ts
@@ -19,6 +19,7 @@ export * from "./huggingface.js";
 export * from "./cloudflareAi.js";
 export * from "./logoSvg.js";
 export * from "./mediaPro.js";
+export * from "./researchBrief.js";
 export * from "./copy.js";
 export * from "./ids.js";
 export * from "./errors.js";